requirements.txt
fastapi[standard]
pillow
